+++ lib/active_interaction/filters/hash_filter.rb
@@ -8,17 +8,13 @@ module ActiveInteraction
     #
     #   @!macro filter_method_params
     #   @param block [Proc] filter methods to apply for select keys
-    #   @option options [Boolean] :strip (true) strip unknown keys (Note: All
-    #     keys are symbolized. Ruby does not GC symbols so this can cause
-    #     memory bloat. Setting this option to `false` and passing in non-safe
-    #     input (e.g. Rails `params`) opens your software to a denial of
-    #     service attack.)
+    #   @option options [Boolean] :strip (true) remove unknown keys
     #
     #   @example
     #     hash :order
     #   @example
     #     hash :order do
-    #       model :item
+    #       object :item
     #       integer :quantity, default: 1
     #     end
   end
@@ -32,11 +28,12 @@ module ActiveInteraction
     def cast(value)
       case value
       when Hash
-        value = stringify_the_symbol_keys(value)
+        value = value.with_indifferent_access
+        initial = strip? ? ActiveSupport::HashWithIndifferentAccess.new : value
 
-        filters.each_with_object(strip? ? {} : value) do |(name, filter), h|
+        filters.each_with_object(initial) do |(name, filter), h|
           clean_value(h, name.to_s, filter, value)
-        end.symbolize_keys
+        end
       else
         super
       end
@@ -78,10 +75,6 @@ module ActiveInteraction
     # @return [Boolean]
     def strip?
       options.fetch(:strip, true)
-    end
-
-    def stringify_the_symbol_keys(hash)
-      self.class.transform_keys(hash) { |k| k.is_a?(Symbol) ? k.to_s : k }
     end
   end
 end