+++ spec/active_interaction/integration/hash_interaction_spec.rb
@@ -21,11 +21,11 @@ describe HashInteraction do
     before { inputs.merge!(a: a) }
 
     it 'returns the correct value for :a' do
-      expect(result[:a]).to eql a.symbolize_keys
+      expect(result[:a]).to eql a.with_indifferent_access
     end
 
     it 'returns the correct value for :b' do
-      expect(result[:b]).to eql(x: {})
+      expect(result[:b]).to eql('x' => {})
     end
   end
 
@@ -36,6 +36,16 @@ describe HashInteraction do
           hash :a, default: Object.new
         end
       end.to raise_error ActiveInteraction::InvalidDefaultError
+    end
+  end
+
+  context 'with an invalid default as a proc' do
+    it 'does not raise an error' do
+      expect do
+        Class.new(ActiveInteraction::Base) do
+          array :a, default: -> { Object.new }
+        end
+      end.to_not raise_error
     end
   end