+++ spec/active_interaction/integration/object_interaction_spec.rb
@@ -1,0 +1,16 @@
+# coding: utf-8
+
+require 'spec_helper'
+
+ObjectInteraction = Class.new(TestInteraction) do
+  object :object
+end
+
+describe ObjectInteraction do
+  include_context 'interactions'
+  it_behaves_like 'an interaction', :object, -> { // }, class: Regexp
+
+  it 'succeeds when given nil' do
+    expect { result }.to_not raise_error
+  end
+end